explain features, misc fixes
	* sample_features: better describe added features in terms of
	  information changed.
	
	* socket.c: fix a compiler warning.

	* README.releasing: add method to release to sourceforge.

+++ Vanilla/ntserv/socket.c
@@ -1325,7 +1325,7 @@ static void handlePlanReq(struct planet_cpacket *packet)
     if (!F_check_planets) return;
     if (packet->pnum < 0 || packet->pnum > MAXPLANETS) return;
 
-    plan = &planets[packet->pnum];
+    plan = &planets[(int) packet->pnum];
 
     if (plan->pl_info & me->p_team) {
         if ( plan->pl_info != packet->info
